Fill out overlapping emptiness; Correctly offset cursor when min/max-aligning blocks

+++ algorithms.h
@@ -6,8 +6,8 @@
 struct algorithm_data {
   struct num_list *numbers;
 
-  int unit_len;
-  int skip_threshold;
+  size_t unit_len;
+  size_t skip_threshold;
 
   char *const row;
   char *const prepare;
